+++ java/client/src/test/java/glide/api/models/TransactionTests.java
@@ -39,6 +39,7 @@ import static redis_request.RedisRequestOuterClass.RequestType.Expire;
 import static redis_request.RedisRequestOuterClass.RequestType.ExpireAt;
 import static redis_request.RedisRequestOuterClass.RequestType.FlushAll;
 import static redis_request.RedisRequestOuterClass.RequestType.GeoAdd;
+import static redis_request.RedisRequestOuterClass.RequestType.GeoPos;
 import static redis_request.RedisRequestOuterClass.RequestType.GetRange;
 import static redis_request.RedisRequestOuterClass.RequestType.GetString;
 import static redis_request.RedisRequestOuterClass.RequestType.HLen;
@@ -123,6 +124,7 @@ import static redis_request.RedisRequestOuterClass.RequestType.ZRemRangeByScore;
 import static redis_request.RedisRequestOuterClass.RequestType.ZRevRank;
 import static redis_request.RedisRequestOuterClass.RequestType.ZScore;
 import static redis_request.RedisRequestOuterClass.RequestType.ZUnion;
+import static redis_request.RedisRequestOuterClass.RequestType.ZUnionStore;
 import static redis_request.RedisRequestOuterClass.RequestType.Zadd;
 import static redis_request.RedisRequestOuterClass.RequestType.Zcard;
 import static redis_request.RedisRequestOuterClass.RequestType.Zcount;
@@ -494,6 +496,9 @@ public class TransactionTests {
         transaction.zinterstore("destination", new KeyArray(new String[] {"key1", "key2"}));
         results.add(Pair.of(ZInterStore, buildArgs("destination", "2", "key1", "key2")));
 
+        transaction.zunionstore("destination", new KeyArray(new String[] {"key1", "key2"}));
+        results.add(Pair.of(ZUnionStore, buildArgs("destination", "2", "key1", "key2")));
+
         transaction.zunion(new KeyArray(new String[] {"key1", "key2"}));
         results.add(Pair.of(ZUnion, buildArgs("2", "key1", "key2")));
 
@@ -508,6 +513,21 @@ public class TransactionTests {
         results.add(
                 Pair.of(
                         ZInterStore,
+                        buildArgs(
+                                "destination",
+                                "2",
+                                "key1",
+                                "key2",
+                                WEIGHTS_REDIS_API,
+                                "10.0",
+                                "20.0",
+                                AGGREGATE_REDIS_API,
+                                Aggregate.MAX.toString())));
+
+        transaction.zunionstore("destination", new WeightedKeys(weightedKeys), Aggregate.MAX);
+        results.add(
+                Pair.of(
+                        ZUnionStore,
                         buildArgs(
                                 "destination",
                                 "2",
@@ -680,6 +700,8 @@ public class TransactionTests {
                                 "10.0",
                                 "20.0",
                                 "Place")));
+        transaction.geopos("key", new String[] {"Place"});
+        results.add(Pair.of(GeoPos, buildArgs("key", "Place")));
 
         var protobufTransaction = transaction.getProtobufTransaction().build();